Kappa_i is always a vector. The promote=true option in DataFrames.jl is used when you want to automatically change the type of  a column to accommodate a new value that doesn't  match the current type of the column.
In your case, the error message suggests that
you're trying to add a value to the :Kappa_i
column that doesn't match its current type.
If you pass promote=true when adding the value,
DataFrames.jl will try to find a common type that
can accommodate both the current values in the
column and the new value you're trying to add.

However, in your current code, you're using the
Agents.jl package to run a model and collect data,
and it doesn't look like it supports the
promote=true option. The run! function in Agents.jl
calls collect_agent_data! to collect data from
the agents, and collect_agent_data! doesn't have
a promote argument.